Clean up long double misuse

+++ test/PalindromeTests.cpp
@@ -280,7 +280,7 @@ public:
 		_l = get<2>(GetParam());
 		_r = get<3>(GetParam());
 	}
-	long double MaxSizePalindromeCountTest()
+	long MaxSizePalindromeCountTest()
 	{
 		return _palindrome.MaxSizePalindromeCount(_data, _l, _r);
 	}
@@ -315,18 +315,16 @@ INSTANTIATE_TEST_SUITE_P(
 					  make_tuple(922261429, "cstniwwvbkyrxzvjpegpgtwwxkdujwbmsqrmkurdprzfftazyonxmawydyjgmipyassxnafluvaouoiuxrqrbrjmzisptfhqqaxq", 19, 81),
 					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 0),
 					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 1),
-					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 2)
-	));
+					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 2)));
 #else
 INSTANTIATE_TEST_SUITE_P(
 	MaxSizePalindromeCountTests,
 	MaxSizePalindromeCountTestFixture,
 	::testing::Values(make_tuple(2, "week", 0, 3),
-		make_tuple(1, "week", 1, 2),
-		make_tuple(2, "abab", 0, 3),
-		make_tuple(1, "wuhmbspjnfviogqzldrcxtaeyk", 20, 20),
-		make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 0),
-		make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 1),
-		make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 2)
-	));
+					  make_tuple(1, "week", 1, 2),
+					  make_tuple(2, "abab", 0, 3),
+					  make_tuple(1, "wuhmbspjnfviogqzldrcxtaeyk", 20, 20),
+					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 0),
+					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 1),
+					  make_tuple(1, "jjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjj", 0, 2)));
 #endif
